### 2019-11-13 #1887 @kbarnhart
@TomasBeuzen if you have recommendations for reviewers please tag them on this thread.

@csherwood-usgs, @dbuscombe-usgs would you be interested in reviewing this paper for the Journal of Open Source Software? The review process is a bit different than other academic journals. The streamlined [review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) takes place on GitHub and focuses on the software and a [very short paper](https://github.com/openjournals/joss-papers/blob/joss.01887/joss.01887/10.21105.joss.01887.pdf).

This here is a pre-review issue. Once I've attracted sufficient reviewers I'll open a dedicated review issue where check-boxes guide you through the review process.
Let me know if you have questions.
