### 2019-11-13 #1887 @csherwood-usgs
I'd be happy to review the python package. The other reviewers tagged are
also very good recommendations.
-Chris

On Tue, Nov 12, 2019 at 8:06 PM Tomas Beuzen <notifications@github.com>
wrote:

> Other potential reviewers: @ebgoldstein <https://github.com/ebgoldstein>,
> @simmonsja <https://github.com/simmonsja>, @ncohn
> <https://github.com/ncohn>, @edlazarus <https://github.com/edlazarus>
>
