### 2019-11-13 #1890 @whedon
[REVIEW]: pybeach: A Python package for extracting the location of dune toes on beach profile transects
**Submitting author:** @TomasBeuzen (<a href="http://orcid.org/0000-0003-2762-9151">Tomas Beuzen</a>)
**Repository:** <a href="https://github.com/TomasBeuzen/pybeach" target ="_blank">https://github.com/TomasBeuzen/pybeach</a>
**Version:** v0.1.1
