### 2019-11-13 #1887 @simmonsja
Likewise a little late and I don't meet the COI criteria for Tom anyhow. Please feel free to consider me for JOSS reviews in the future!
