### 2019-11-14 #1887 @kbarnhart
@dbuscombe-usgs and @simmonsja thanks for being willing to review in general. Consider adding yourself to the list of people willing to review for JOSS [here](https://joss.theoj.org/reviewer-signup.html).
