### 2019-11-25 #1890 @csherwood-usgs
Dear Bot and Katy,
Thanks for the opportunity to review this useful contribution. I have
submitted some issues to the repo and checked all of the boxes on the
review, but there is no indication that my part of the process is complete,
so I am sending this email into the void to let you know.
Regards, Chris
