### 2019-11-26 480a09a
add dependencies section
