### 2019-11-28 #1890 @kbarnhart
@csherwood-usgs thanks for making your review as issues in the main repo. I've linked them here (just by including a hyperlink). This helps create the record of what issues and/or changes in the main repo were made as part of this review.

@ncohn @edlazarus let me know if you have any questions as you proceed with your reviews. If you make any issues in the main repo, just link them with this issue using a markdown hyperlink
```
[display this](url_goes_here)
```

Once you've all finished your reviews and @TomasBeuzen has finished addressing them, I'll ask each of you to affirm your recommendation to accept the package for publication in JOSS.

I'll be OOO Dec 2nd-5th with no internet (GMT+12).
