### 2019-12-02 #1890 @ncohn
Echoing much of what Chris and Eli have said - this is a nice contribution that is likely to benefit many in the coastal process field. I personally expect to use this for my work since having a rapid approach for characterizing the dune toe via multiple approaches is incredibly useful.

As part of this review, I tested against profiles from the U.S. Pacific Northwest (PNW) which the commonly used maximum curvature approach rarely does a good job of for estimating a realistic dune toe position. Unsurprisingly, the relative relief and maximum curvature approaches coded into “pybeach” often give drastically different estimates for the dune toe for PNW dunes – just as you would expect from those methods. I was quite impressed with the new ML approach that was included, which generally matched the “eye estimated” dune toe well. I did not do an extensive assessment of the tool for all types of dune morphologies, but on my first assessment I am quite impressed. Nice job Tom!

Overall, the python code is well written and is easy to use for a relatively novice python user. The JOSS paper is also well written, concise, and gives an appropriate overview of the tool. Chris noted some dependency issues that I also ran into on my first attempt with the tool, but seem to have been fixed with updates to the repo. As I continue to use the code I will add issues on the git-hub repo, but at this stage I don’t see any need major issues that require attention before publication.
