### 2019-12-05 #1890 @kbarnhart
Thanks to @csherwood-usgs @ncohn @edlazarus for your reviews.

@TomasBeuzen, please ping me here when you've finished addressing reviewer comments.
