### 2019-12-06 #1890 @TomasBeuzen
@kbarnhart - I believe I've addressed all reviewer comments.

@csherwood-usgs @ncohn @edlazarus - thank you so much for your time and effort in reviewing pybeach, it's very much appreciated!
