### 2019-12-07 #1890 @kbarnhart
@TomasBeuzen can you close the issues in your repository that you have finished addressing? I may have inadvertently re-opened one of them while linking with this repo (sorry).

@csherwood-usgs @ncohn @edlazarus if you are satisfied with the changes made in response to your reviews, can you please confirm you recommendation to accept this submission by commenting here (this is a standard part of the JOSS process).
