### 2019-12-07 TomasBeuzen/pybeach#7 @kbarnhart
Paper edits for JOSS
This issue summarizes a few topical editor requests for the content of the paper [under review at JOSS](https://github.com/openjournals/joss-reviews/issues/1890).

@TomasBeuzen can you

- [x] Provide a very short definition of "beach" and "dune" in the context of shore profile geometry. I request this to make the summary more understandable for a diverse, non-specialist audience.

- [x] Specify you are using the Random Forest classifier in the summary rather than referring to the more general "Machine Learning". I think this may be as simple as stating "Machine learning using Random Forest classification- discussed further below." in your enumerated list of supported methods.

- [x] When you are discussing support for custom ML models from user supplied data, can you be more specific about which ML methods are supported. Random forest with other data? Anything SciKit-learn supports?

- [x] Your paper nicely describes your extensive work to compare the performance of the alternative methods provided in **pybeach**. However, this is lost a bit in the **pybeach** section. I'd recommend that you add a new section heading called something like "Performance Assessment" just before the sentences that start with "For each dataset described above, the true location of the dune toe..."
