### 2019-12-10 TomasBeuzen/pybeach#7 @TomasBeuzen
Thank you @kbarnhart for your great comments - all have now been addressed! To be clear:

1. New text has been added to briefly clarify to non-specialist audiences the "beach" and "dune": _"Sandy coastlines typically comprise two key parts: a beach and dune. The beach is the section of sandy coast that is mostly above water (depending upon tide) and actively influenced by waves, while dunes are elevated mounds/ridges of sand at the back of the beach. The interface between the beach and dune is often characterised by a distinct change in ground slope (with the dune having a steeper slope than the beach)."_

2. Have now referred to "Machine learning using Random Forest classification" as per your suggestion.

3. Have now clarified that pybeach supports any scikit-learn model that supports probabilistic prediction.

4. Have added a new section heading "Performance assessment" at the appropriate location.
