### 2019-12-16 #1890 @kbarnhart
@TomasBeuzen great! Thanks for the ping (I was sucked into the AGUniverse last week).

Can you
- [ ] Fix the [Harley2017] citation (sorry I missed these two citation related commenst in my earlier issue).
- [ ] Add a date to Leaman (n.d.)
- [ ] Make a tagged release and archive. Ensure that the title and author list are the same as the paper.
- [ ] Report the current version number and archive DOI in the here.

At that point I will be able to update the version and archive DOI. I will then be able to hand the publication over for final processing by the Editors in Chief.

Nice work! 🎉
