### 2019-12-16 7225eeb
fix reference to Harley2017
