+++ paper.bib
@@ -83,7 +83,8 @@
   title={chrisleaman/py-wave-runup: v0.1.4},
   author={Chris Leaman},
   doi = {http://doi.org/10.5281/zenodo.2697004},
-  journal={Zenodo}
+  journal={Zenodo},
+  year={2019}
 }
 
 @article{Beuzen2019,
