### 2019-12-16 #1890 @TomasBeuzen
Hi @kbarnhart - no problem! I unfortunately missed AGU this year but it looked great!

- Harley2017 and Leaman2019 citations have been corrected.
- Have made a tagged release and archived with Zenodo.
- Version number: v0.1.1
- DOI: 10.5281/zenodo.3579500 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3579501.svg)](https://doi.org/10.5281/zenodo.3579501)
