### 2019-12-16 #1890 @kbarnhart
@openjournals/joss-eics this is ready for final processing.
