### 2019-12-20 #1890 @arfon
@TomasBeuzen - I spotted a typo in your paper here: https://github.com/TomasBeuzen/pybeach/pull/8 . Also, I think you mention scikit-learn in the paper - perhaps a [citation](https://scikit-learn.org/stable/about.html#citing-scikit-learn) would be appropriate?
