### 2019-12-20 cd23d6e
add sklearn citation
