+++ paper.bib
@@ -109,3 +109,16 @@
   year={2013},
   publisher={Wiley Online Library}
 }
+
+@article{sklearn2011,
+ title={Scikit-learn: Machine Learning in {P}ython},
+ author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
+ and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
+ and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
+ Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
+ doi={},
+ journal= {Journal of Machine Learning Research},
+ volume= {12},
+ pages= {2825--2830},
+ year= {2011}
+}
