### 2019-12-20 10d9cf2
include additional scikit-learn citation
