### 2019-12-20 #1890 @TomasBeuzen
@arfon - thanks for the typo fix and yes definitely agree a scikit-learn citation is necessary, sorry I missed that!

Paper has now been updated.
