### 2019-12-20 #1890 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
